Delat upp koden från main till nya mappar. Rensat i utkommenterad kod.
