+++ lib/ProfileTab.dart
@@ -1,0 +1,274 @@
+import 'dart:convert';
+
+import 'package:cached_network_image/cached_network_image.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_tindercard/flutter_tindercard.dart';
+import 'package:http/http.dart' as http;
+import 'package:rflutter_alert/rflutter_alert.dart';
+import 'package:stycling/Profile/EditProfile.dart';
+import 'package:stycling/settings.dart';
+
+import 'Profile/ProfilePage.dart';
+import 'settings.dart';
+
+class ProfileTab extends StatefulWidget {
+  String CurrentUserID;
+
+  ProfileTab(String CurrentUserID) {
+    this.CurrentUserID = CurrentUserID;
+  }
+
+  @override
+  _ProfileTabState createState() => _ProfileTabState();
+}
+
+class _ProfileTabState extends State<ProfileTab> {
+  @override
+  void dispose() {
+    super.dispose();
+  }
+
+  Future<List<dynamic>> getMyPlants(String url) async {
+    String url =
+        "https://sticklingar.herokuapp.com/plants/" + widget.CurrentUserID;
+    final response = await http.get(url);
+
+    final responseJson = json.decode(response.body.toString());
+
+    List<dynamic> plants = (json.decode(response.body) as List);
+
+    return plants;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: SingleChildScrollView(
+        child: Container(
+          child: Column(
+            children: <Widget>[
+              Padding(
+                padding: EdgeInsets.only(top: 15, left: 10),
+                child: Row(
+                  children: <Widget>[
+                    IconButton(
+                      icon: Icon(Icons.settings),
+                      color: Colors.grey,
+                      onPressed: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                              builder: (context) => SettingsPage()),
+                        );
+                      },
+                      iconSize: 40,
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.only(left: 50.0),
+                      child: Text(
+                        'My Sticklings',
+                        style: TextStyle(
+                            fontSize: 30,
+                            fontFamily: 'Lato',
+                            fontWeight: FontWeight.bold),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              Padding(
+                padding: EdgeInsets.only(top: 30),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    GestureDetector(
+                      child: RaisedButton(
+                        onPressed: null,
+                        padding: EdgeInsets.all(15),
+                        shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(20)),
+                        child: Text(
+                          'Add new plant',
+                          style: TextStyle(
+                            fontSize: 25,
+                            //fontWeight: FontWeight.bold,
+                            color: Colors.white,
+                            fontFamily: 'Lato',
+                          ),
+                        ),
+                        disabledColor: Color(0xFF65C27A),
+                      ),
+                      onTap: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                              builder: (context) =>
+                                  PlantProfile(null, widget.CurrentUserID)),
+                        );
+                      },
+                    )
+                  ],
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.only(top: 28.0),
+                child: FutureBuilder(
+                  future: getMyPlants(""),
+                  builder: (context, snapshot) {
+                    if (!snapshot.hasData) {
+                      return Center(
+                          child: CircularProgressIndicator(
+                        backgroundColor: Colors.teal,
+                      ));
+                    } else {
+                      List plantImages = snapshot.data;
+
+                      return Container(
+                        height: 500,
+                        child: ListView.builder(
+                          itemCount: plantImages.length,
+                          itemBuilder: (context, i) {
+                            return Padding(
+                              padding: const EdgeInsets.all(10.0),
+                              child: Container(
+                                height: 100,
+                                width: 150,
+                                color: Colors.white,
+                                child: Row(
+                                  children: [
+                                    ClipRRect(
+                                      borderRadius: BorderRadius.circular(20.0),
+                                      child: CachedNetworkImage(
+                                        imageUrl: plantImages[i]["pic"],
+                                        placeholder: (context, url) => SizedBox(
+                                          width: 80,
+                                          height: 80,
+                                          child: new CircularProgressIndicator(
+                                            backgroundColor: Colors.teal,
+                                          ),
+                                        ),
+                                        errorWidget: (context, url, error) =>
+                                            new Icon(Icons.error),
+                                      ),
+                                    ),
+                                    Expanded(
+                                      child: Column(
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.start,
+                                        crossAxisAlignment:
+                                            CrossAxisAlignment.start,
+                                        children: <Widget>[
+                                          Padding(
+                                            padding: const EdgeInsets.only(
+                                              bottom: 8.0,
+                                              right: 10,
+                                              left: 20,
+                                            ),
+                                            child: Row(
+                                              mainAxisAlignment:
+                                                  MainAxisAlignment
+                                                      .spaceBetween,
+                                              crossAxisAlignment:
+                                                  CrossAxisAlignment.start,
+                                              children: [
+                                                Text(
+                                                  plantImages[i]["title"],
+                                                  style: TextStyle(
+                                                      fontFamily: 'Lato',
+                                                      fontWeight:
+                                                          FontWeight.bold,
+                                                      color: Colors.black,
+                                                      fontSize: 20),
+                                                ),
+                                                GestureDetector(
+                                                  onTap: () {
+                                                    Navigator.push(
+                                                        context,
+                                                        MaterialPageRoute(
+                                                            builder: (context) =>
+                                                                EditProfile(
+                                                                    plantImages[
+                                                                            i]
+                                                                        ["pic"],
+                                                                    plantImages[
+                                                                            i][
+                                                                        "title"],
+                                                                    plantImages[
+                                                                            i][
+                                                                        "desc"])));
+                                                  },
+                                                  child: Container(
+                                                    width: 90,
+                                                    height: 35,
+                                                    decoration: BoxDecoration(
+                                                        border: Border.all(
+                                                          color:
+                                                              Colors.grey[500],
+                                                        ),
+                                                        borderRadius:
+                                                            BorderRadius.all(
+                                                                Radius.circular(
+                                                                    20))),
+                                                    child: Row(
+                                                      mainAxisAlignment:
+                                                          MainAxisAlignment
+                                                              .center,
+                                                      children: [
+                                                        Icon(
+                                                            Icons.edit_outlined,
+                                                            color: Colors.grey,
+                                                            size: 25),
+                                                        Text(
+                                                          "Edit",
+                                                          style: TextStyle(
+                                                              fontFamily:
+                                                                  'Lato',
+                                                              fontWeight:
+                                                                  FontWeight
+                                                                      .normal,
+                                                              color:
+                                                                  Colors.black,
+                                                              fontSize: 20),
+                                                        ),
+                                                      ],
+                                                    ),
+                                                  ),
+                                                )
+                                              ],
+                                            ),
+                                          ),
+                                          Padding(
+                                            padding: EdgeInsets.only(left: 20),
+                                            child: Text(
+                                              plantImages[i]["desc"],
+                                              maxLines: 2,
+                                              style: TextStyle(
+                                                  fontFamily: 'Lato',
+                                                  fontWeight: FontWeight.normal,
+                                                  color: Colors.black,
+                                                  fontSize: 20),
+                                            ),
+                                          ),
+                                        ],
+                                      ),
+                                    )
+                                  ],
+                                ),
+                              ),
+                            );
+                          },
+                        ),
+                      );
+                    }
+                  },
+                ),
+              )
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
